# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GymGest API",
        "description": "Gym management system API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1 Server"
        }
    ],
    "paths": {
...
        "/api/v1/gyms/{gymId}/students": {
            "get": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Listar alumnos de un gimnasio",
                "operationId": "38de82bef58500a734ca2610788c9b85",
                "parameters": [
                    {
                        "name": "gymId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alumnos"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Gimnasio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Matricular alumno en un gimnasio",
                "operationId": "1fd931e3cc88a49f3a8746fb31d5810b",
                "parameters": [
                    {
                        "name": "gymId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "quota_expires_at"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "quota_expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alumno matriculado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Email no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/personal-training/students": {
            "post": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Matricular alumno en entrenamiento personal",
                "operationId": "d5ac0dc9e000a33d5f99dbbc5359908a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "quota_expires_at"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "quota_expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alumno matriculado en entrenamiento personal"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Email no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gyms/{gymId}/students/{studentId}": {
            "put": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Modificar cuota de un alumno",
                "operationId": "3db25c531e14379a4fc7dd9e769dc16f",
                "parameters": [
                    {
                        "name": "gymId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quota_expires_at"
                                ],
                                "properties": {
                                    "quota_expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cuota actualizada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Dar de baja a un alumno (baja lógica)",
                "operationId": "d394c7424a0cb43d4fd47ea8c1c6784a",
                "parameters": [
                    {
                        "name": "gymId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Alumno dado de baja"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gyms/{gymId}/students/{studentId}/reactivate": {
            "put": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Dar de alta a un alumno inactivo",
                "operationId": "c3f8705d64f768180fce4b4019cee28d",
                "parameters": [
                    {
                        "name": "gymId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "studentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quota_expires_at"
                                ],
                                "properties": {
                                    "quota_expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alumno reactivado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/students": {
            "get": {
                "tags": [
                    "Gym Students"
                ],
                "summary": "Listar todos los alumnos de todos los gimnasios del entrenador",
                "operationId": "8f763de7be2e28bc042aa889853bd6ae",
                "responses": {
                    "200": {
                        "description": "Lista de todos los alumnos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}